Move StatusRow to _output
